fix: get correct endpoint class during registration
When adding security annotation to an endpoint, Quarkus generates a subclass
to handle method security.
EndpointRegistry tries to detect the base class using Spring utility class,
but only CGLIB proxies are considered.
This change replaces the Spring ClassUtils.getUserClass call in EndpointRegistry
with an implementation that discards all synthetic classes