fix: exclude offending classes in the extension (#21)
Move offending classes exclusion from the jandex index
to the extension, so in a future the extension can be
based on an official index.
The 'removedResources' configuration causes the jandex index to be rebuilt,
but the quarkus-hilla artifact does not contain classes,
resulting in an empty index.
For this reason a new index has to be rebuilt against hilla artifacts,
therefore the removal configuration needs to be duplicated,
to prevent the usage of offending classes.

Co-authored-by: Dario Götze <Dario@ok.de>