chore: help IntelliJ IDEA to handle hilla-shaded-deps module (#60)
IntelliJ IDEA fails to handle the hilla-shaded-deps module, causing compilation
errors to be shown in the IDE.
This change adds a maven configuration to attach the same artifact to the maven
project with a custom classifier, so that IDEA is able to correctly find the
classes. However, the plugin execution is always skipped, so it does not affect
the build process.