chore(actions): set validation status check for ignored paths
If a PR contains only changes to files ignored by the validation
workflow, the build and test job is not executed and its status is
pending, preventing the PR to being merged.

This change introduces an additional workflow that marks as good PR that
contains changes to ignored files.

See https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-protected-branches/troubleshooting-required-status-checks\#handling-skipped-but-required-checks
for more information.