test: prevent reusing dev-server
Quarkus tests reuse the same application instance, unless tests define
different profiles or test resources. However, the JVM is always the same
and Flow reuses the existing dev server instance.
This can cause errors when the application is restarted if frontend files
are forcibly deleted.
This change make sure that the dev server is always stopped and restarted.