test: fix confirm dialog verifications
Vaadin switched to native popover for dialog components.
However, it looks like selenium does not handle correctly
visibility of elements in a popover since they are actually
visible but their opacity is 1, resulting in test failures
when the element is clicked.
This change introduces a workaround to click a button in a
popover by using Selenium Actions API.