test: use ordering to run E2E tests affecting database
TestTransaction annotation does not work for E2E tests because backend
services do not use the same transaction started by the annotation interceptor.
This change introduces ordering for test executions so that tests that
modify data are executed after read-only tests to somplify assertions.
A better solution would be to reset the database at every execution,
but it would slow down the overall tests execution.