fix: refactor Vaadin ServiceInitEvent propagation (#1849)
Removes the QuarkusVaadinServiceListenerPropagator class and the build
step that creates the META-INF/services file for that
VaadinServiceInitListener because it caused issue with uber-jar
deployment by overwriting entries provided by other jars.
The change adds an observer method for Vaadin ServiceInitEvent and
forwards it to all `@VaadinServiceEnabled` bean implementing
VaadinServiceInitListener interface.
This is a temporary workaround, waiting for a proper fix in
vaadin-quarkus extension (see vaadin/quarkus#234)