fix: remove workaround for VaadinServiceInitListener beans
Removes the QuarkusVaadinServiceListenerPropagator class and the build
step that creates the META-INF/services file for that
VaadinServiceInitListener because it caused issue with uber-jar
deployment by overwriting entries provided by other jars.
'VaadinServiceInitListener's are now lookeup up correcty by
vaadin-quarkus (since 2.2.3), so the propagator is not needed anymore.
The change also adds an observer method for Vaadin ServiceInitEvent to
forward the event to all `@VaadinServiceEnabled` bean implementing
VaadinServiceInitListener interface, used only for previous versions of
vaadin-quarkus..